Screenshot Page Redo & CSS changes
Redid the screenshots page. Now the page table is wider and the
screenshots larger. Too much detail was lost in the scaled down
pictures, especially for larger networks with small labels and neurons.
This also gave more room for text.

Some minor cropping was done on a few of the screenshots.

Some CSS was taken from the documentation CSS and put on the website
CSS since things like headers weren’t stylistically formatted. This
also creates more of a homogeneity of appearance between the docs and
the website, which is desirable.